+++ utils/calculate_distance.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   calculate_distance.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ribana-b <ribana-b@student.42malaga.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/18 21:26:27 by ribana-b          #+#    #+# Malaga      */
+/*   Updated: 2025/01/18 22:04:47 by ribana-b         ###   ########.com      */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "utils.h"
 
-float	calculate_distance(t_player *player, t_map *map, float angle)
+double	calculate_distance(t_player *player, t_map *map, double angle)
 {
-	float	distance;
+	double	distance;
 	t_v2	start;
 	t_v2	end;
