+++ utils/v2/v2mult.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   v2mult.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ribana-b <ribana-b@student.42malaga.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/18 21:25:07 by ribana-b          #+#    #+# Malaga      */
+/*   Updated: 2025/01/18 21:25:19 by ribana-b         ###   ########.com      */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "utils.h"
 
 t_v2	v2_mult(t_v2 *v1, t_v2 *v2)
@@ -8,7 +20,7 @@ t_v2	v2_mult(t_v2 *v1, t_v2 *v2)
 	});
 }
 
-t_v2	v2_mult_scalar(t_v2 *v1, float scalar)
+t_v2	v2_mult_scalar(t_v2 *v1, double scalar)
 {
 	return ((t_v2){
 		.x = v1->x * scalar,
